pyproject.toml
[tool.poetry]
name = "ocelot"
version = "0.1.0"
description = ""
authors = ["Ranieri Althoff <ranisalt@gmail.com>"]
packages = [{ include = "ocelot" }]

[tool.poetry.dependencies]
cryptography = "^38.0.4"
pydantic = "^1.10.2"
pyjwt = { extras = ["crypto"], version = "^2.6.0" }
python = "^3.10"
starlette = "^0.21.0"
toml = "^0.10.2"
tortoise-orm = { extras = ["asyncmy"], version = "^0.19.2" }
uvicorn = { extras = ["standard"], version = "^0.18.3" }

[tool.poetry.group.dev.dependencies]
asgi-lifespan = "^1.0.1"
black = { allow-prereleases = true, version = "*" }
Faker = "^15.3.4"
httpx = "^0.23.1"
pytest = "^7.2.0"
pytest-benchmark = "^4.0.0"
pytest-cov = "^4.0.0"
tortoise-orm = { extras = ["aiosqlite"], version = "^0.19.2" }
types-toml = "^0.10.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
